chore: move docs-site to apps/ with pnpm, oxfmt, and TypeScript 7
Adopt a pnpm workspace under apps/docs-site, pin JavaScript dependencies
exactly (including TypeScript 7.0.2), and format Markdown/JSON/YAML plus
the docs app from the repo root with oxfmt while ignoring generated output.
