+++ .github/workflows/codeql.yml
@@ -8,7 +8,7 @@ on:
       - main
   # Weekly run at 09:00 UTC on Friday (18:00 JST Friday)
   schedule:
-    - cron: '0 9 * * 5'
+    - cron: "0 9 * * 5"
 permissions:
   contents: read
 jobs:
