+++ apps/docs-site/.gitignore
