+++ apps/docs-site/content/plugins/compression.mdx
