+++ apps/docs-site/content/plugins/meta.json
@@ -1,0 +1,4 @@
+{
+  "title": "Plugins",
+  "pages": ["index", "serialization", "connect-get", "body-limit", "compression", "protovalidate", "code-generator"]
+}
