+++ apps/docs-site/waku.config.ts
@@ -6,9 +6,9 @@ import mdx from "fumadocs-mdx/vite";
 const isProd = process.env.NODE_ENV === "production";
 
 export default defineConfig({
-  // Project Pages: https://ichizero.github.io/connect-ktor/
-  basePath: isProd ? "/connect-ktor/" : "/",
-  vite: {
-    plugins: [press(), mdx(), tailwindcss()],
-  },
+    // Project Pages: https://ichizero.github.io/connect-ktor/
+    basePath: isProd ? "/connect-ktor/" : "/",
+    vite: {
+        plugins: [press(), mdx(), tailwindcss()],
+    },
 });
