+++ config/detekt/detekt.yml
@@ -49,4 +49,3 @@ exceptions:
       - "**/src/test/**"
       - "**/src/testFixtures/**"
       - "**/conformance/**"
-
