--- docs-site/README.md
@@ -1,19 +1,0 @@
-# Connect-Ktor documentation site
-
-Fumapress (Fumadocs + Waku) site published to GitHub Pages at
-https://ichizero.github.io/connect-ktor/.
-
-## Scripts
-
-```bash
-npm install
-npm run dev          # local preview (basePath `/`)
-npm run build        # static export to dist/public (basePath `/connect-ktor/`)
-npm run types:check
-```
-
-From the repository root:
-
-```bash
-mise run docs-site:build
-```
