--- docs-site/content/plugins/meta.json
@@ -1,12 +1,0 @@
-{
-  "title": "Plugins",
-  "pages": [
-    "index",
-    "serialization",
-    "connect-get",
-    "body-limit",
-    "compression",
-    "protovalidate",
-    "code-generator"
-  ]
-}
