--- docs-site/source.config.ts
@@ -1,16 +1,0 @@
-import { defineDocs } from "fumadocs-mdx/config";
-import { metaSchema, pageSchema } from "fumapress/adapters/mdx/schema";
-
-export const docs = defineDocs({
-  dir: "content",
-  docs: {
-    async: true,
-    schema: pageSchema,
-    postprocess: {
-      includeProcessedMarkdown: true,
-    },
-  },
-  meta: {
-    schema: metaSchema,
-  },
-});
