+++ gradle/libs.versions.toml
@@ -54,8 +54,8 @@ okhttp = { module = "com.squareup.okhttp3:okhttp", version.ref = "okhttp" }
 slf4j-simple = { module = "org.slf4j:slf4j-simple", version.ref = "slf4j" }
 
 junit-jupiter-engine = { module = "org.junit.jupiter:junit-jupiter-engine", version.ref = "junit-jupiter-engine" }
-kotest-runner-junit5 = { module = "io.kotest:kotest-runner-junit5", version.ref = "kotest"}
-kotest-framework-datatest = { module = "io.kotest:kotest-framework-datatest", version.ref = "kotest"}
+kotest-runner-junit5 = { module = "io.kotest:kotest-runner-junit5", version.ref = "kotest" }
+kotest-framework-datatest = { module = "io.kotest:kotest-framework-datatest", version.ref = "kotest" }
 kotest-assertions-core = { module = "io.kotest:kotest-assertions-core", version.ref = "kotest" }
 kotest-assertions-json = { module = "io.kotest:kotest-assertions-json", version.ref = "kotest" }
 kotlinx-serialization = { module = "org.jetbrains.kotlinx:kotlinx-serialization-json", version.ref = "kotlinx-serialization" }
@@ -65,10 +65,7 @@ spotless = { module = "com.diffplug.spotless:spotless-plugin-gradle", version.re
 detekt-plugin = { module = "io.gitlab.arturbosch.detekt:detekt-gradle-plugin", version.ref = "detekt" }
 
 [bundles]
-connect = [
-    "connect-kotlin",
-    "connect-kotlin-google-java-ext",
-]
+connect = ["connect-kotlin", "connect-kotlin-google-java-ext"]
 ktor = [
     "ktor-server-core-jvm",
     "ktor-server-compression",
@@ -77,12 +74,7 @@ ktor = [
     "ktor-server-status-pages",
     "ktor-server-body-limit",
 ]
-protobuf = [
-    "protobuf-java",
-    "protobuf-java-util",
-    "protobuf-kotlin",
-    "protovalidate",
-]
+protobuf = ["protobuf-java", "protobuf-java-util", "protobuf-kotlin", "protovalidate"]
 test = [
     "junit-jupiter-engine",
     "kotest-runner-junit5",
