+++ package.json
@@ -1,0 +1,15 @@
+{
+  "name": "connect-ktor",
+  "private": true,
+  "scripts": {
+    "fmt": "oxfmt",
+    "fmt:check": "oxfmt --check",
+    "docs:dev": "pnpm --filter docs-site dev",
+    "docs:build": "pnpm --filter docs-site build",
+    "docs:types": "pnpm --filter docs-site types:check"
+  },
+  "devDependencies": {
+    "oxfmt": "0.61.0"
+  },
+  "packageManager": "pnpm@10.14.0"
+}
